+++ engine/hid/src/hid_private.h
@@ -84,7 +84,7 @@ namespace dmHID
     // TODO: start using the dmUser namespace
     // TODO: How to represent user id from/to C/Lua
     //      I.e. convert from uint32_t to platform type
-    bool GetPlatformGamepadUserId(HContext context, HGamepad gamepad, uint32_t* user_id);
+    bool GetPlatformGamepadUserId(HContext context, HGamepad gamepad, uint16_t* user_id);
     int  GetKeyValue(Key key);
     int  GetMouseButtonValue(MouseButton button);
 }
